requirements.txt
python-binance==1.0.19
pandas>=2.3.2
pandas-ta==0.4.71b0
python-dotenv==1.0.1
